Refactor locking logic into its own class so we can use it elsewhere
