+++ vscode-dotnet-runtime-library/src/Acquisition/InstallTrackerSingleton.ts
@@ -39,6 +39,7 @@ import { DotnetCoreAcquisitionWorker } from './DotnetCoreAcquisitionWorker';
 import { IEventStream } from '../EventStream/EventStream';
 import { IExtensionState } from '../IExtensionState';
 import { IDotnetAcquireContext } from '..';
+import { executeWithLock } from './LockUtilities';
 
 interface InProgressInstall
 {
@@ -77,54 +78,6 @@ export class InstallTrackerSingleton
         InstallTrackerSingleton.instance.extensionState = extensionState;
     }
 
-    protected executeWithLock = async <A extends any[], R>(alreadyHoldingLock : boolean, dataKey : string, f: (...args: A) => R, ...args: A): Promise<R> =>
-    {
-        const trackingLock = `${dataKey}.lock`;
-        const lockPath = path.join(__dirname, trackingLock);
-        fs.writeFileSync(lockPath, '', 'utf-8');
-
-        let returnResult : any;
-
-        try
-        {
-            if(alreadyHoldingLock)
-            {
-                // eslint-disable-next-line @typescript-eslint/await-thenable
-                return await f(...(args));
-            }
-            else
-            {
-                this.eventStream?.post(new DotnetLockAttemptingAcquireEvent(`Lock Acquisition request to begin.`, new Date().toISOString(), lockPath, lockPath));
-                await lockfile.lock(lockPath, { retries: { retries: 10, minTimeout: 5, maxTimeout: 10000 } })
-                .then(async (release) =>
-                {
-                    // eslint-disable-next-line @typescript-eslint/await-thenable
-                    returnResult = await f(...(args));
-                    this.eventStream?.post(new DotnetLockReleasedEvent(`Lock about to be released.`, new Date().toISOString(), lockPath, lockPath));
-                    return release();
-                })
-                .catch((e : Error) =>
-                {
-                    // Either the lock could not be acquired or releasing it failed
-                    this.eventStream?.post(new DotnetLockErrorEvent(e, e.message, new Date().toISOString(), lockPath, lockPath));
-                });
-            }
-        }
-        catch(e : any)
-        {
-            // Either the lock could not be acquired or releasing it failed
-
-            // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-            this.eventStream.post(new DotnetLockErrorEvent(e, e?.message ?? 'Unable to acquire lock to update installation state', new Date().toISOString(), lockPath, lockPath));
-
-            // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-            throw new EventBasedError('DotnetLockErrorEvent', e?.message, e?.stack);
-        }
-
-        return returnResult;
-    }
 
     public clearPromises() : void
     {
@@ -176,7 +129,7 @@ export class InstallTrackerSingleton
 
     public async canUninstall(isFinishedInstall : boolean, dotnetInstall : DotnetInstall, allowUninstallUserOnlyInstall = false) : Promise<boolean>
     {
-        return this.executeWithLock( false, this.installedVersionsId, async (id: string, install: DotnetInstall) =>
+        return executeWithLock( this.eventStream, false, this.installedVersionsId, async (id: string, install: DotnetInstall) =>
         {
             this.eventStream.post(new RemovingVersionFromExtensionState(`Removing ${JSON.stringify(install)} with id ${id} from the state.`));
             const existingInstalls = await this.getExistingInstalls(id === this.installedVersionsId, true);
@@ -189,7 +142,7 @@ export class InstallTrackerSingleton
 
     public async uninstallAllRecords() : Promise<void>
     {
-        await this.executeWithLock( false, this.installingVersionsId, async () =>
+        await executeWithLock( this.eventStream, false, this.installingVersionsId, async () =>
         {
             // This does not uninstall global things yet, so don't remove their ids.
             const installingVersions = await this.getExistingInstalls(false, true);
@@ -197,7 +150,7 @@ export class InstallTrackerSingleton
             await this.extensionState.update(this.installingVersionsId, remainingInstallingVersions);
         }, );
 
-        return this.executeWithLock( false, this.installedVersionsId, async () =>
+        return executeWithLock( this.eventStream, false, this.installedVersionsId, async () =>
         {
             const installedVersions = await this.getExistingInstalls(true, true);
             const remainingInstalledVersions = installedVersions.filter(x => x.dotnetInstall.isGlobal);
@@ -211,7 +164,7 @@ export class InstallTrackerSingleton
      */
     public async getExistingInstalls(getAlreadyInstalledVersion : boolean, alreadyHoldingLock = false) : Promise<InstallRecord[]>
     {
-        return this.executeWithLock( alreadyHoldingLock, getAlreadyInstalledVersion ? this.installedVersionsId : this.installingVersionsId,
+        return executeWithLock( this.eventStream, alreadyHoldingLock, getAlreadyInstalledVersion ? this.installedVersionsId : this.installingVersionsId,
             (getAlreadyInstalledVersions : boolean) =>
         {
             const extensionStateAccessor = getAlreadyInstalledVersions ? this.installedVersionsId : this.installingVersionsId;
@@ -284,7 +237,7 @@ ${convertedInstalls.map(x => `${JSON.stringify(x.dotnetInstall)} owned by ${x.in
 
     protected async removeVersionFromExtensionState(context : IAcquisitionWorkerContext, idStr: string, installIdObj: DotnetInstall, forceUninstall = false)
     {
-        return this.executeWithLock( false, idStr, async (id: string, install: DotnetInstall) =>
+        return executeWithLock( this.eventStream, false, idStr, async (id: string, install: DotnetInstall) =>
         {
             this.eventStream.post(new RemovingVersionFromExtensionState(`Removing ${JSON.stringify(install)} with id ${id} from the state.`));
             const existingInstalls = await this.getExistingInstalls(id === this.installedVersionsId, true);
@@ -332,7 +285,7 @@ ${installRecord.map(x => `${x.installingExtensions.join(' ')} ${JSON.stringify(I
 
     protected async addVersionToExtensionState(context : IAcquisitionWorkerContext, idStr: string, installObj: DotnetInstall, alreadyHoldingLock = false)
     {
-        return this.executeWithLock( alreadyHoldingLock, idStr, async (id: string, install: DotnetInstall) =>
+        return executeWithLock( this.eventStream, alreadyHoldingLock, idStr, async (id: string, install: DotnetInstall) =>
         {
             this.eventStream.post(new RemovingVersionFromExtensionState(`Adding ${JSON.stringify(install)} with id ${id} from the state.`));
 
@@ -369,7 +322,7 @@ ${existingVersions.map(x => `${JSON.stringify(x.dotnetInstall)} owned by ${x.ins
 
     public async checkForUnrecordedLocalSDKSuccessfulInstall(context : IAcquisitionWorkerContext, dotnetInstallDirectory: string, installedInstallIdsList: InstallRecord[]): Promise<InstallRecord[]>
     {
-        return this.executeWithLock( false, this.installedVersionsId, async (dotnetInstallDir: string, installedInstallIds: InstallRecord[]) =>
+        return executeWithLock( this.eventStream, false, this.installedVersionsId, async (dotnetInstallDir: string, installedInstallIds: InstallRecord[]) =>
         {
             let localSDKDirectoryIdIter = '';
             try
